requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
ansi2html==1.9.1
anyio==4.2.0
asgiref==3.7.2
astroid==3.0.2
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.2
blinker==1.7.0
bs4==0.0.1
build==1.0.3
cachetools==5.3.2
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.3.29
click==8.1.7
clickhouse-connect==0.7.0
coloredlogs==15.0.1
colorhash==2.0.0
dataclasses-json==0.6.3
Deprecated==1.2.14
dill==0.3.7
distro==1.9.0
docx==0.2.4
duckdb==0.9.2
emoji==2.9.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
extra-streamlit-components==0.1.60
faiss-cpu==1.7.4
fastapi==0.85.1
filelock==3.13.1
filetype==1.2.0
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.26.2
googleapis-common-protos==1.62.0
grandalf==0.8
greenlet==3.0.3
grpcio==1.60.0
h11==0.14.0
hnswlib==0.8.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.2
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
iniconfig==2.0.0
install==1.3.5
isodate==0.6.1
isort==5.13.2
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kubernetes==29.0.0
langchain==0.1.11
langchain-community==0.0.26
langchain-core==0.1.29
langchain-openai==0.0.5
langchain-text-splitters==0.0.1
langdetect==1.0.9
langsmith==0.1.22
loguru==0.7.2
lxml==5.1.0
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.2
mccabe==0.7.0
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
neo4j==5.16.0
networkx==3.2.1
nltk==3.8.1
numexpr==2.8.8
numpy==1.26.3
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.11.1
openpyxl==3.1.2
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
orjson==3.9.15
overrides==7.4.0
packaging==23.2
pandas==2.1.4
pillow==10.2.0
platformdirs==4.1.0
pluggy==1.3.0
posthog==3.3.1
protobuf==4.25.2
pulsar-client==3.4.0
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.14
pydantic_core==2.14.6
pydeck==0.8.0
Pygments==2.17.2
PyJWT==2.8.0
pylint==3.0.3
Pympler==1.0.1
pyparsing==3.1.1
pypdf==4.0.0
PyPika==0.48.9
pyproject_hooks==1.0.0
pytest==7.4.4
pytest-env==1.1.3
pytest-mock==3.12.0
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
python-iso639==2024.1.2
python-magic==0.4.27
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
rapidfuzz==3.6.1
rdflib==7.0.0
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
starlette==0.20.4
streamlit==1.31.1
streamlit-agraph==0.0.45
streamlit-authenticator==0.3.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.2
tokenizers==0.15.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.3
toolz==0.12.0
torch==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
transformers==4.36.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
unstructured==0.12.0
unstructured-client==0.15.2
urllib3==2.1.0
uvicorn==0.25.0
uvloop==0.19.0
validators==0.22.0
watchdog==4.0.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
wikipedia==1.4.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
zstandard==0.22.0
